Don't store async boolean in nodes
The place argument to transform() basically means "please put this
statement into this array", and the return value of transform() means
"successive statements should go into this array". When the node is
asynchronous, the two value differ (most likely, the argument is the
containing block and the return value is the body of a callback
function). So there's no need to remember that separately, comparing
the original block body and the transform() return value is enough.

This also fixes some cases where transform() wasn't called, its
return value was ignored, or async wasn't properly being set.